Create a simple web page stub
This commit adds a basic stub for building a web page. The page itself
is just a title with the usual `Hello World!`, but this commit already
contains some interesting ideas:
- there is no JavaScript but TypeScript. But instead of using the type-
  script compiler to detect all modules, it will be invoked on a per-
  file basis.
- the `ninja`-based build system compiles the TypeScript files to an
  output directory. The other resources are copied over as well.

This implies that one cannot run the HTTP server inside the repository
root or even the `src` directory. Instead a server has to be started in
the `build` directory after invoking `ninja`.