Add a service worker and PWA manifest
This commit adds the core components for the PWA at last. There is a so-
called manifest, which contains some of the core settings such as the
color and images and how the PWA should be started.
There is also a so-called service worker, which gets registered by the
main JavaScript code. Note, that this service worker does not yet cache
anything but simply passes requests through to the actual server.