Allow building a WASM file from Rust
This commit supports building a WASM file, i.e. WebAssembly, from Rust
source code. The old workflow using ninja is kept, i.e. in order to
build the whole site, you still only need to invoke `ninja` (no need for
manually calling `cargo` or similar).

The current Rust source is only very minor: the `lib.rs` contains only a
single function, which is not even properly exported. Note, that this
commit uses a non-standard Rust layout: the `Cargo.toml` (and also the
`Cargo.lock`) file is in the repository root, while the source is not
in `src`, but in `src/wasm` in order to play nicer with the other source
files like TypeScript.
The `Cargo.lock`-file is tracked as well to ensure reproducible builds.

One thing about the two build-systems at play: there is now both `ninja`
and `cargo`. This still uses cargo, but `ninja` calls it, whenever ne-
cessary. The dependency-tracking is done using the `.d`-file generated
by rustc. The ninja-directives `depfile` and `deps` ensure, that this
file (containing all the GCC-style dependencies) is read and deleted, so
that, on any changes in the Rust source code, `cargo` is re-executed.
By specifying the `deps = gcc` the d-file is deleted, so that no unne-
cessary rebuild is done.

The Rust code is always compiled using `--release`.