Install `wasm-bindgen` CLI
This is a CLI required for generating the JS-bindings to the Rust code
yet to be written. In order to build the software later on, this CLI has
to be added to the docker image.
While doing this, the Rust install step is put in front of the other
software installations, so that this `RUN`-step can be cached and the
build is faster.