Use `tsc --outDir` instead of `--outFile`
This is required in order to properly use modules in the TypeScript
source code. The `$outDir` is `$site/scripts` by default, i.e. each .ts
script will be compiled into the `/scripts` directory. Since the service
worker should be located in the root of the server, its `--outDir` is
overwritten.