Add reproducible builds for PNGs
This commit addresses an issue in the `gh-pages`-branch: on different
deploys, the same `icon.svg` is rendered to the various PNGs. Somewhat
unexpectedly, the same input yields different outputs, so that the re-
pository (the `gh-pages` branch) contains multiple times the same PNGs,
just with different timestamps.
Therefore this commit excludes those timestamps from the generated PNGs.