Render a full-screen X onto a canvas
This commit uses `web-sys` to insert a canvas into the `<body>` and is
using the 2D rendering context to render a full-screen X onto that new
canvas. This way a simply rendering can be shown.

Note, that this commit does not use `winit` but does the rendering-
"loop" all by hand (currently only when resizing). In the future, one
may want to migrate to winit for easier cross-platform application or
generally easier development.