Use `winit` for abstracting the I/O handling
The popular `winit` crate allows to create a so-called event loop, which
is a cross-platform way for receiving "OS"-events. The OS, in this case,
is the web browser.
Note, that this depends on a `beta`-release, as this release already
contains many changes for the used "web" target and hence prevents mi-
gration issues later on.

Note, that `winit` leaves the actual rendering up to the user. Therefore
this commit sticks to the 2D-canvas-rendering.