Clean up the canvas creation
This commit moves the DOM manipulation (i.e. creating the canvas as well
as a container and appending both to the `<body>`) into a separate
module `dom` to encapsulate it, as it is relatively decoupled from any
other rendering stuff.
This also removes the intermediate rendering of the X onto the canvas.
Instead, all `winit` events are now logged.