Dynamically resize the canvas to the maximum size
This commit uses a [`ResizeObserver`] to get notified when the parent
container of the canvas gets resized. This happens, for example, when
the browser window is resized.
When the `ResizeObserver` is triggered, the canvas is resized to match
the size of the parent `<div>`.

`winit` also [has the same mechanism][pr], which listens for changes of
the canvas itself in order to report the new size through the corres-
ponding `WindowEvent::Resized`-event. This way the user code receives
the new size.

[`ResizeObserver`]: https://developer.mozilla.org/en-US/docs/Web/API/ResizeObserver
[pr]: https://github.com/rust-windowing/winit/pull/2859/files#diff-da3475798eee382f6f9bb4d8f1310a5adcbd606ec8080ee0cb5f4497bc0f9a01