Support `<canvas>` resizing on mobile
This commit is interesting: before it, mobile user only see a blank
screen because the canvas is either not properly sized or gets resized
to the wrong values.
There are two reason for this, one only involves CSS and the other needs
changes in the Rust code:
- one needs to specify the widths and heights of the canvas to support
  the proper size (else the canvas "collapses"). This happens e.g. in
  mobile firefox in landscape orientation
- the code only considered a typical desktop viewport e.g. in horizontal
  orientation. Previously the code unconditionally used the properties
  `inline_size`/`block_size` as if the viewport is horizontal. Therefore
  the code now needs to check, which `writing-mode` is computed for the
  canvas element and, if it is a vertical one, swap the two dimensions.

After applying this commit, the black X on green ground also renders on
mobile devices.