Display a warning if WebAssembly is unsupported
This commit generalizes the pre-existing `<noscript>`-mechanism and
styling to general "fullscreen-warnings", that totally prevent the usage
of the application. This generalization is done by renaming the classes
in the HTML and CSS files.

The next thing is the addition of a hidden-by-default warning for the
case that WASM is not supported. This is un-hidden only if the initiali-
zation of the WASM threw an exception. If the initialization did work,
it is simply left as-is, i.e. hidden.

The wording of the warning messages is also a little bit harmonized.