Don't create canvas dynamically
Previously the WASM code did create the canvas ands its parent container
all by itself and inserted it into the DOM (at the end of the `<body>`).
While this is generally fine and also prevents issues with getting the
HTML DOM and the WASM code out of sync (as the WASM was independent), it
has a major downside: the canvas cannot be easily embedded in other ele-
ments. Of course one could search for a specific container in the DOM,
still create the canvas programmatically and then insert it in the found
container, it has the same downsides with regards to the two sides begin
required to be in sync. Therefore it is better to just create a full DOM
in the HTML source and query the required items.
This commit does exactly this. While this is not strictly necessary at
the moment, it will be easier to do HTML changes later on.