Make the canvas fill up all possible space
This commit adds a few CSS rules, that ensure, that the container of the
canvas takes up all remaining space in the window (the canvas is resized
to match the container programmatically). This is achieved by using a
flex-parent and rules, that automatically allow growing and shrinking
the child whenever needed. The `min-height`-rule is really important:
without it, the container and canvas are growing but not shrinking.