Add a footer with information about cirq
This commit extends the full-screen-div with a footer, which displays
potentially helpful information such as the link to the source code and
the version (although this version is currently hardcoded to "unknown").

To match the `<footer>`, the rendering-related stuff is now moved into
a `<main>` element, which seems to be a good fit.