Use standard `tsc` for compiling `.ts`-files
This commit uses the standard way of compiling TypeScript sources with
the TypeScript compiler `tsc` by making use of the configuration files.
Since there are issues with service-/web-workers, this again uses a
somewhat special layout of typescript files such as described in this
[stackoverflow] answer.
By using the `tsconfig.json` it is also possible to remove the tripple-
slash-annotations specifying the libraries used.

Note, that the layout in `src` does not correspond 1:1 to the layout in
the `build/site` directory anymore, e.g.: the `worker.js` is in the root
of the site-directory while it is inside the `scripts/service-worker`-
directory inside the source-tree. But doing it this way separates the
different source languages properly and doesn't clutter the root with
`tsconfig`s.

[stackoverflow][https://stackoverflow.com/a/56374158]