Prevent spurious TypeScript rebuilds
This commit adds the `restat`-flag to the `tsc`-invocation. The reason
for this is, that `tsc` only writes files, that actually changed. The
problem is, that `ninja` requires a common rule and therefore expects,
that all outputs are regenerated. Previously this caused rebuilds of one
has build everything, then changes one of the files, e.g. `worker.ts`
and then builds. The `main.js` is not regenerated and ninja mentions:

    ninja -d explain
    ninja explain: output build/tsc/main/main.js older than most recent input src/scripts/service-worker/worker.ts (1693050764019645844 vs 1693051083382966667)
    ninja explain: build/tsc/main/main.js is dirty
    ninja explain: build/tsc/service-worker/worker.js is dirty
    [3/3] Copy build/tsc/service-worker/worker.js to build directory

This commit tells `ninja` to check the output timestamps and accept them
as they are in the moment, preventing this issue.