Support using wasm-bindgen-generated `d.ts`
Up to now, the import of the `wasm.js` file was not resolved by the
TypeScript-compile and hence any type hints, e.g. for `init`, were un-
available. This commit changes it.

Thankfully, `wasm-bindgen` outputs a `.d.ts`-file for the generated JS-
file, so that TypeScript can reason about the types. The main change is
to include the `typeRoots`-configuration for `tsc`, but there are many
additional changes required:
* one needs to refer to the module by the name `"wasm"`
* as the browser does not support loading scripts this way, a new rule
  `support-no-bundler` was added, which replaces the import statements
  `"<name>"` with `"./<name>.js"`.
* as the `.d.ts`-file is now required for the compilation of TypeScript
  files, the generation by `wasm-bindgen` must happen before attempting
  to compile these files. A dependency is added to ensure this.
* while being on it, the new `support-no-bundler`-rule also removes all
  `export`-lines, as they are also not supported inside browsers.