Move main application WASM to subdirectory
This commit moves the files in `src/wasm/` to `src/wasm/app/` in order
to allow further WASM binaries alongside it. The `Cargo.toml` of this
`wasm` crate is moved into that directory as well. In the crate root,
there is a new virtual manifest which holds the profile settings and
contains the paths to the workspace members.