Start a WASM binary for the service worker
This commit lays the foundation for using WASM inside the service worker
(which is currently entirely JavaScript). This way one can write nearly
the whole service worker within Rust.
The actual commit is not that spectacular: it builds upon the last com-
mits, e.g. it uses the modified `cargo`-rule to build another crate in
the root workspace.

Note, that the WASM is not currently used.