Worker stub: use JavaScript instead of TypeScript
This commit converts the `worker.ts` to `worker.js` and removes every
TypeScript-related stuff around it, as well as adopting the building
instructions.
The reason for this is, that the file is a mere stub and does not do any
complicated work, where TypeScript would help. Therefore it is just kind
of an overhead and thus is removed. JavaScript will work sufficiently
here, especially as the file is not expected to be modified anymore (one
should modify the Rust code of the service worker instead).