Mark `tsconfig.tsbuildinfo` as outputs of `tsc`
This way those files are deleted on `ninja -t clean`. Not doing so makes
`tsc` think, that it does not need to do anything, as it looks at the
build-info and "knows", that it has build the outputs and so they should
be there. With this commit, the build-info is cleaned as well, so this
ill-condition cannot happen anymore.