Force whole rebuilds of the TypeScript-project
This is somewhat unfortunate, but necessary for the following reason.
The TypeScript-compiler `tcs` has an internal incremental compilation,
that does not take the type-definitions (`.d.ts`) into account, i.e.
even if the `ninja` triggered a re-compilation, `tsc` will not actually
recomile anything.
This is especially bad if someone changes the signature of a public
function, which causes `wasm-bindgen` to generate changed type defini-
tions and finally `tsc` to _ignore_ the changes altogether... This will
**not** detect the errors in the TypeScript-files!

Fortunately, this is only relevant locally, but thankfully not in CI, as
that always does clean builds.