Don't use a WASM within the service worker anymore
In a perfect world, the service worker would just forward to a WASM
(similar to the main application), but this is currently not that easy
for multiple reasons:

* one needs to register the event handlers on the first execution (i.e. on
  the top-level) of the worker script. This prevents registering the WASM
  doing it altogether.
* TypeScript only plays nice with modules, but [at the moment], Firefox does
  not support service workers as modules, as this feature is experimental.
  Therefore one cannot do simply `import * as wasm from "service_worker"` to
  get the WASM code loaded. The alternative is to use `importScripts()`, but
  then the type annotations are missing for TypeScript. Dynamic imports are
  forbidden by the specification altogether.

In conclusion, there is ([at the moment]) no real benefit for using WASM in
the service worker, when cross-browser-support is required (it is!). So, the
whole service worker is implemented in this single script, which makes it
work in all browsers, that support service workers, regardless of such
experimental features like module-support.

Therefore this commit removes all the WASM-related parts from the build
description. This is very unfortunate, but at the moment, it is the only
sensible thing to do (of course, one could import the WASM-script via
`importScripts()` and ignore, that there is no typing possible and then
forward to plain old Rust functions inside the WASM, but there are no
advantages in doing so).

Fixes #26.

[at the moment]: https://caniuse.com/mdn-api_serviceworker_ecmascript_modules