Fix loading screen animation on slow networks
The loading screen animation previously was not properly started. This
was not a problem, if the network was fast enough (but I don't know why)
and only broke on slower networks, e.g. "Regular 2G" in the Firefox Net-
working-tab.
This commit fixes it by actually playing the animation and removing it
directly afterwards, now, that the "finish" event works (which it did
not previously, as the animation was not properly started).