Use custom title bar on desktop PWAs
This commit makes use of the (experimental) "window-controls-overlay"
feature. While normally one should not use experimental features, this
one is fine, because it is a pure opt-in for browsers, that support it,
i.e. there is no negative impact on browsers not supporting it.

The main difference is, that this hides most of the title bar of the
system and replaces it with the header, that is already there (which
currently only contains the logo and the title). This makes up a much
more natural feeling on the desktop. The whole title bar serves as the
so-called draggable space, just like on the normal operating systems.

This commit already makes use of the non-stable `app-region`-property,
so that it simply works once the feature is stable.

Based on [this article](https://web.dev/window-controls-overlay/).