Center a smaller search bar on large windows
This commit only applies to windows with a minimum width of 1000px (this
value was chosen arbitrarily, but works well at my desktop PWA and my
mobile phone). If the screen is large, expanding the search bar to the
whole width looks ugly. Therefore this commit changes it, so that (on
large screens) the search bar has a suitable maximum width and is in the
exact center of the title bar.