Add a worker interface for the application code
This commit introduces the `WorkerInterface` class, which abstracts away
the interaction details between the application code and the service
worker, e.g. to search for updates.
This class currently implements the search for updates: depending on the
availability of the experimental "periodic sync API" that API or a fall-
back is used to send a request to the service worker. The fallback is
implemented as a periodic `postMessage()`. The main difference is, that
the fallback only runs if the tab is open and needs to run more often.

The service worker needs to handle both kinds of requests (periodic sync
and ordinary message). Currently it only logs the update request, but
doesn't do anything more than that.

Note, that this uses a second module, which needs `export item`. Pre-
viously all export-lines were removed entirely, which was wrong: only
the service worker must not have the additional `export default null;`-
line, that is only necessary for TypeScript anyways. Therefore the build
script is adopted to to only filter out that line for the service worker
and to don't interfere with normal application modules.