Fix the `support-no-bundler`-rule
This rule previously did delete all lines, that contain a line starting
with `export`. This was wrong for multiple reasons: only the service
worker must not have the additional `export default null;`-line, that is
only necessary for TypeScript anyways. Another important point is, that
the main application code might need to use another module at some point
and this module will need to export classes/functions/etc. which would
be broken by this rule.

Therefore the build script is adopted to to only filter out that line
for the service worker and to don't interfere with normal application
modules.