Insert the version number into the service worker
This commit introduces a version number in the service worker, that is
increased on each software build. This is the recommended way to com-
municate changes in the application version to the (offline) PWA: the
browser sees, that the content of the service worker has changed and in-
stalls a new one (at that point the service worker can fetch new ver-
sions of all files). MDN explicitly mentions this approach [here].

Note, that this diverges from the previous planned way on doing updates!

This commit uses the ninja-build-system to automatically process the
service worker source and replace `$version` with the content of the
`.version`-file created by the last commit. As the special `parameter`
for the `support-no-bundler` is not used anymore, it is removed.

[here]: https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps/Tutorials/CycleTracker/Service_workers#version_number