Add the list of static resources to service worker
This list can be generated with the command given in the description.
Note, that it would be good to generate this list at compile-time, but
this is not easily possible: this rule would need to run at the very
last step in the `ninja`-invocation. But this would need to include all
other files as (order-only-)dependencies, which just moves the complexi-
ty from the easy-to-understand TypeScript-file to the harder-to-under-
stand build-definition-file. Therefore the solution described in the
commit is easiest and thus fine for the moment.