Implement the caching of all site resources
This commit implements the offline-only-strategy in a very simple and
straight-forward way. There are no unexpected steps and everything is
very similar to the [MDN tutorial].
With this commit all resources are cached on installation and are served
from this cache on any fetch (requests not cached are responded with a
custom 404 response). The new version is only activated once the tab/PWA
is closed and reopended (or a force-reload happens), i.e. there is no
call to `skipWaiting()`.

[MDN tutorial]: https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps/Tutorials/CycleTracker/Service_workers#version_number