Add way to update the worker on user's behalf
This commit adds a way to activate a service worker if a new one is
available. On a reload of the `main.js`, the browser attempts to re-
download the service worker and, if there is a difference, it installs
the new one. The service worker is not activated automatically after
the installation, though, in order to allow the user to select, when to
update.
If a new service worker is available (which currently only happens after
a manual reload), the user gets a new (currently unstyled) button to
perform the update. If the button is not clicked, nothing happens, but
on a click the service worker gets sent a message to activate itself and
serve all clients from now on. Refer to the ASCII diagram for details.

The next steps are to display the alert in a more user-friendly fashion
and to search for updates in the background.