Make the whole titlebar draggable
The titlebar is artificially enlarged on desktop PWAs in order to fit
the search bar into the titlebar. The problem was, that only the header-
element was marked as draggable, but that element does not span 100% of
the window width, but is only in between of the overlay elements (the
system buttons). This prevented dragging the window at the space below
those UI elements.
Therefore this commit adds a new div `.titlebar`, which only is as large
as the usable space, but the previous `<header>`-element takes now 100%
of the width. The whole header is draggable, so the problem is fixed.

Furthermore the new titlebar-container is now positioned absolute, which
is required on platforms, that have the system UI on the left side.